+++ packages/@molt/command/src/Input/index.ts
@@ -1,0 +1,1 @@
+export * as Input from './Input.js'
