+++ packages/@molt/command/src/builder.ts
@@ -1,0 +1,38 @@
+import type { FlagSpecExpressionParseResultToPropertyName } from './helpers.js'
+import { getLowerCaseEnvironment } from './helpers.js'
+import { Input } from './Input/index.js'
+import { Settings } from './Settings/index.js'
+import type { FlagName } from '@molt/types'
+import type { Any } from 'ts-toolbelt'
+import type { z } from 'zod'
+
+// prettier-ignore
+type ParametersToArguments<ParametersSchema extends z.ZodRawShape> = Any.Compute<{
+  [FlagSpecExpression in keyof ParametersSchema as FlagSpecExpressionParseResultToPropertyName<FlagName.Parse<FlagSpecExpression & string>>]:
+    z.infer<ParametersSchema[FlagSpecExpression]>
+}>
+
+type Definition<ParametersSchema extends z.ZodRawShape> = {
+  parseOrThrow: (processArguments?: string[]) => ParametersToArguments<ParametersSchema>
+  settings: (newSettings: Settings.Input<ParametersSchema>) => Definition<ParametersSchema>
+  schema: ParametersSchema
+}
+
+export const create = <Schema extends z.ZodRawShape>(schema: Schema): Definition<Schema> => {
+  const settings = {
+    ...Settings.getDefaults(getLowerCaseEnvironment()),
+  }
+
+  const api = {
+    settings: (newSettings) => {
+      Settings.change(settings, newSettings)
+      return api
+    },
+    parseOrThrow: (processArguments) => {
+      // eslint-disable-next-line
+      return Input.parseOrThrow(schema, processArguments ?? process.argv.slice(2), settings) as any
+    },
+    schema,
+  } as Definition<Schema>
+  return api
+}
